# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Evidencija Recepata API",
        "description": "REST API za upravljanje receptima, sastojcima i korisnicima.",
        "contact": {
            "email": "admin@evidencijarecepata.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Glavni server"
        }
    ],
    "paths": {
...
        "/api/sastojci": {
            "post": {
                "tags": [
                    "Sastojci"
                ],
                "summary": "Kreiranje novog sastojka",
                "operationId": "93295e54c381d3c633f0fb6a1b13dc09",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ime",
                                    "kolicina",
                                    "recept_id"
                                ],
                                "properties": {
                                    "ime": {
                                        "type": "string",
                                        "example": "Brašno"
                                    },
                                    "kolicina": {
                                        "type": "string",
                                        "example": "500g"
                                    },
                                    "recept_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "dostupno": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sastojak kreiran"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    },
                    "401": {
                        "description": "Neautorizovan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/sastojci/{id}": {
            "delete": {
                "tags": [
                    "Sastojci"
                ],
                "summary": "Brisanje sastojka",
                "operationId": "e728b4c5d7fe6a0f76942d9efec31318",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sastojak obrisan"
                    },
                    "404": {
                        "description": "Sastojak nije pronađen"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/sastojci/{id}/dostupno": {
            "put": {
                "tags": [
                    "Sastojci"
                ],
                "summary": "Izmena dostupnosti sastojka (samo dobavljač)",
                "operationId": "6b6a2f1ed44f81a82e5a7d6a13ddf024",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "dostupno"
                                ],
                                "properties": {
                                    "dostupno": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Dostupnost izmenjena"
                    },
                    "403": {
                        "description": "Zabranjen pristup"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Unesi token dobijen pri registraciji/prijavi",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}